webhooks: show 20 deliveries by default, add --num flag
The previously hard-coded number of 1000 puts strain on the webhooks
API subsystem and can cause requests to fail. While that's not ideal,
there's also no way to use a different number to work around the
issue.

1000 is probably a bit much for normal use, too, since it can easily
scroll most results off the screen. 20 is used as the default,
matching `heroku releases`.

The new --num flag, also matching `heroku releases`, can be used to
specify a different number.